Moved types.js
