+++ client/src/components/Admin/AdminLayout.jsx
@@ -1,7 +1,7 @@
 import { Component } from 'react';
 import { withRouter } from 'react-router-dom';
 import { arrayOf, func, string, shape } from 'prop-types';
-import { Location, Match, Size } from '../../types';
+import { Location, Match, Size } from '../common/types';
 import { isUpToMedium } from '../../styles/utils';
 
 /* eslint-disable arrow-body-style */
