+++ client/src/components/common/SnackBar/SnackBar.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Snackbar from 'material-ui/Snackbar';
 import { bool, func, number, string } from 'prop-types';
 import { $blueCool, $red } from '../../../styles/variables';
-import { Intent } from '../../../types';
+import { Intent } from '../types';
 
 const successStyle = {
   backgroundColor: $blueCool
